government/requirements.txt
asgiref==3.2.3
Django==3.0
freeze==1.0.10
pycrypto==2.6.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
